refactoring-update-user

+++ isps/src/app/plans/plansSA.js
@@ -2,11 +2,15 @@
 
 import { db } from "@/app/lib/turso";
 import { nanoid } from "nanoid";
+import { getUser } from "../lib/getUser";
 
 
 export async function plansServerAction(_, formData) {
 
-   
+    const currentUser = await getUser();
+    if (!currentUser?.id) return { ok: false, message: "You must be logged in" };
+
+    const adminId = currentUser.id;
 
     const planId = formData.get("public_id")?.toString().trim();
 
@@ -20,8 +24,8 @@ export async function plansServerAction(_, formData) {
 
         if (!planId && speed && rate) {
             const result = await db.execute(
-                `INSERT INTO plans (public_id, speed, rate) VALUES (?, ?, ?)`,
-                [nanoid(), speed, rate]
+                `INSERT INTO plans (public_id, speed, rate, admin_id) VALUES (?, ?, ?, ?)`,
+                [nanoid(), speed, rate, adminId]
             );
 
             if (result.rowsAffected === 0) {
@@ -33,8 +37,8 @@ export async function plansServerAction(_, formData) {
 
         if (planId && speed && rate) {
             const result = await db.execute(
-                `UPDATE plans SET speed = ?, rate = ? WHERE public_id = ?`,
-                [speed, rate, planId]
+                `UPDATE plans SET speed = ?, rate = ? WHERE public_id = ? AND admin_id = ?`,
+                [speed, rate, planId, adminId]
             );
 
             if (result.rowsAffected === 0) {
@@ -46,8 +50,8 @@ export async function plansServerAction(_, formData) {
 
         if (planId && !speed && !rate) {
             const result = await db.execute(
-                `DELETE FROM plans WHERE public_id = ?`,
-                [planId]
+                `DELETE FROM plans WHERE public_id = ? AND admin_id = ?`,
+                [planId, adminId]
             );
 
             if (result.rowsAffected === 0) {
